starting writing serial processor test
